Refactor: Improve code organization and state management
